Moving server public interfaces
